omnicalc_params_3

+++ app/controllers/calc_controller.rb
@@ -11,8 +11,8 @@ class CalcController < ApplicationController
   end
   def flex_payment
     @apr = params.fetch("apr").to_f/100
-    @num_yrs = params.fetch("num_yrs").to_f
-    @principal = params.fetch("principal").to_f
+    @num_yrs = params.fetch("num_yrs").to_i
+    @principal = params.fetch("principal").to_i
     
     render("calc_templates/flexible_payment.html.erb")
   end
